updates for Windows and file management
some updates for Windows specific code, as well as updates for meflib changes: if the CHANNEL struct is passed in, then these functions should not modify whether or not to close the data files (but should open them if they are not open already).  If the CHANNEL struct is not passed in, then it is created, files are opened, and should always be closed when the function ends.